Refactored some methods from pod-helper
