+++ src/lib/pod-helper.js
@@ -1,5 +1,6 @@
 const fc = require("solid-file-client")
 const MESSAGE_FILE = "messages.json"
+const txtFileBuilder = require("./textFileBuilder.js")
 
 class PODHelper{
 	
@@ -21,66 +22,13 @@ class PODHelper{
 	    return fc.createFile(friendRoute, message).then(200);
     }
 	
-	/**
-	*Generate an ACL text string that grants owner all permissions and Read only to partnerID
-	*@param {String} partnerID
-	*@return {String} ACL string content
-	*/
-	generateACL(partnerID) {
-		partnerID = partnerID.replace("#me", "#");
-		var ACL = "@prefix : <#>. \n"
-			+"@prefix n0: <http://www.w3.org/ns/auth/acl#>. \n"
-			+"@prefix c: </profile/card#>. \n"
-			+"@prefix c0: <"+ partnerID + ">. \n\n"
-
-			+":ControlReadWrite \n"
-				+"\ta n0:Authorization; \n"
-				+"\tn0:accessTo <"+ MESSAGE_FILE +">; \n"
-				+"\tn0:agent c:me; \n"
-				+"\tn0:mode n0:Control, n0:Read, n0:Write. \n"
-			+":Read \n"
-				+"\ta n0:Authorization; \n"
-				+"\tn0:accessTo <"+ MESSAGE_FILE +">; \n"
-				+"\tn0:agent c0:me; \n"
-				+"\tn0:mode n0:Read.";
-			
-		return ACL;
-	}
-	
 	grantReadPermissionsToFile(fileRoute, partnerID) {
 		var aclRoute = fileRoute+".acl";
-		var aclContents = this.generateACL(partnerID);
+		var aclContents = txtFileBuilder.generateACL(partnerID, MESSAGE_FILE);
 		console.log("A 404 ERROR NEXT MEANS PERMISSIONS FILE HAS BEEN SUCCESFULLY CREATED");
 		fc.updateFile(aclRoute, aclContents).then(success => {
 			console.log("Permissions successfully configured")
 		}, err => fc.createFile(aclRoute, aclContents).then(200));
-	}
-	
-	/*
-	* Builds a JSON format file that contains all messages
-	* @param {String} senderID
-	* @param {String} receiverID
-	* @param {Array} messages
-	* @return {String} desired JSON-format contents in a string
-	*/
-	buildJSONmessages(senderID, receiverID, messages) {
-		var sender = senderID.replace("https://", "").replace("/profile/card#me", "");
-		var receiver = receiverID.replace("https://", "").replace("/profile/card#me", "");
-		var lastupdate = new Date().getTime();
-		
-		var jsonstring = JSON.stringify({"webid_sender":sender, "webid_receiver":receiver, "lastupdate":lastupdate, "messages":"////"});
-		var segments = jsonstring.split("////");
-		var ret = "////"+segments[0]+"[";
-		var i;
-		for (i=0; i<messages.length; i++) {
-			ret = ret+messages[i].serialize();
-			if (i!=messages.length-1)
-				ret=ret+",";
-			else 
-				ret = ret+"]"+segments[1]+"////";
-		}
-		alert(ret);
-		return ret;		
 	}
 	
 	/**
@@ -106,7 +54,7 @@ class PODHelper{
 		console.log("A 404 ERROR NEXT MEANS FOLDER HAS BEEN SUCCESFULLY CREATED");
 		fc.createFolder(folderRoute).then(200);
 		
-		var messagesJSON = this.buildJSONmessages(userID, partnerID, messages);
+		var messagesJSON = txtFileBuilder.buildJSONmessages(userID, partnerID, messages);
 		
 		console.log("A 404 ERROR NEXT MEANS MESSAGE LOG FILE HAS BEEN SUCCESFULLY CREATED");
 		fc.updateFile(podFileRoute, messagesJSON).then(success => {
